Restore google analytics
With the site refresh, main.js was no longer included.
Moved the analytics logic to a separate file, and
include it again.
