+++ _includes/footer.html
@@ -51,5 +51,6 @@
 
     <!-- Custom javascript -->
     <script src="{{ site.baseurl }}/resources/js/functions.js" type="text/javascript"></script>
+    <script src="{{ site.baseurl }}/resources/js/analytics.js" type="text/javascript"></script>
   </body>
 </html>
